+++ test/test_IFD_enlargement.c
@@ -108,7 +108,7 @@ static int write_image_data(TIFF *tif, uint16_t width, uint16_t length,
                        TIFFmin(bufLen, (size_t)lastlinebytesmax));
             }
             /* Change something for next row. */
-            pbufLine[0] = (unsigned char)((pbufLine[0] + 35) % bpsmod);
+            pbufLine[0] = (unsigned char)(((unsigned int)pbufLine[0] + 35) % bpsmod);
         }
     }
     else
@@ -128,7 +128,7 @@ static int write_image_data(TIFF *tif, uint16_t width, uint16_t length,
                        TIFFmin(bufLen, (size_t)lastlinebytesmax));
             }
             /* Change something for next row. */
-            pbufLine[0] = (unsigned char)((pbufLine[0] + 30) % bpsmod);
+            pbufLine[0] = (unsigned char)(((unsigned int)pbufLine[0] + 30) % bpsmod);
         }
     }
     _TIFFfree(pbufLine);
@@ -847,7 +847,7 @@ static int test_IFD_enlargement(const char *filename, unsigned int openMode,
                 goto failure;
             }
             bytesRead =
-                TIFFReadScanline(tif, &bufLine[NUMIFDsMAX][0], length - 1, 0);
+                TIFFReadScanline(tif, &bufLine[NUMIFDsMAX][0], (uint32_t)(length - 1), 0);
         }
         if (bytesRead > 0)
         {
