+++ tools/fax2ps.c
@@ -366,7 +366,7 @@ int main(int argc, char **argv)
             case 'p': /* print specific page */
                 pageNumber = (uint16_t)atoi(optarg);
                 if (pages)
-                    pages = (uint16_t *)realloc(pages, (npages + 1) *
+                    pages = (uint16_t *)realloc(pages, (size_t)(npages + 1) *
                                                            sizeof(uint16_t));
                 else
                     pages = (uint16_t *)malloc(sizeof(uint16_t));
